use size_t
git-svn-id: http://svn.xp-dev.com/svn/wxextension@4928 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/include/wx/extension/report/frame.h
@@ -125,7 +125,7 @@ protected:
   void OnIdle(wxIdleEvent& event);
 private:
   void CreateDialogs();
-  void DoRecent(wxFileHistory& history, int index, long flags = 0);
+  void DoRecent(wxFileHistory& history, size_t index, long flags = 0);
   void FindInFiles(wxWindowID dialogid);
   void UseHistory(wxWindowID id, wxMenu* menu, wxFileHistory& history);
